+++ client/src/components/boards-list/constants.ts
@@ -1,6 +1,4 @@
 import { makeStyles } from '@material-ui/core';
-import { SchemaOf } from 'yup';
-import * as yup from 'yup';
 
 const useStyles = makeStyles((theme) => ({
   cardContainer: {
@@ -30,20 +28,4 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-type Validation = {
-  [x: string]: string;
-};
-
-const configValidationSchema = (fieldName: string): SchemaOf<Validation> =>
-  yup.object({
-    [fieldName]: yup
-      .string()
-      .strict()
-      .trim('New name cannot include leading and trailing spaces')
-      .min(2, 'New name must be more than 2 symbols')
-      .max(30, 'Max length is 30 symbols')
-      .required('New name is required')
-      .matches(/^(?:[A-Za-z]+)(?:[A-Za-z0-9 _]*)$/, 'New name must have numbers and letters')
-  });
-
-export { useStyles, configValidationSchema };
+export { useStyles };
