+++ client/utils/get-token.ts
@@ -1,8 +1,8 @@
 import store from '../src/store/store';
 
-function getToken(): string {
+const getToken = (): string => {
   const state = store.getState();
   return state.authData.token;
-}
+};
 
 export default getToken;
