--- client/utils/name-capitalized.ts
@@ -1,5 +1,0 @@
-function nameCapitalized(string: string): string {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-export default nameCapitalized;
